Add .gitignore
added gitignore and reinstalled the dependencies for socket.io and socket.io-client in both the /frontend and /messagingserver directories.

Will need Moesif CORS Changer extension to test out on google chrome, needs to be turned on. If you create a duplicate tab on Chrome, you can talk to yoursefl within the same chat.